+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # OSGL Tool Change Log
 
-1.11.4 11/May/2018
+1.12.0 13/May/2018
+* Allow it to change a mapping stage's semantic #91
+* **compatible break change** Make `rootClass` specified in `Lang.fieldsOf()` be exclusive #90
 * Allow waive global filter for a mapping process #89
 * `Lang.bool("false")` returns `true` #88
 * Add `Lang.isPrimitiveType(Class)` and `Lang.isWrapperType(Class)` method #87
 * Add `Lang.isPrimitiveType(String)` method #86
+
+1.11.4 11/May/2018
 * Mapping global filter update #85
 
 1.11.3 10/May/2018